refactor(connect): move nostr method definitions to corresponding folder
