+++ packages/connect-common/src/types/api/nostr/nostrGetPublicKey.ts
@@ -1,6 +1,6 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { ExperimentalMethod, GetPublicKey, Params, Response } from '../params';
+import type { ExperimentalMethod, GetPublicKey, Params, Response } from '../../params';
 
 export declare function nostrGetPublicKey(
     params: Params<GetPublicKey & ExperimentalMethod>,
