+++ packages/connect-common/src/types/api/nostr/nostrSignEvent.ts
@@ -1,7 +1,7 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { ExperimentalMethod, Params, Response } from '../params';
-import type { NostrSignEvent } from './nostr';
+import type { NostrSignEvent } from './common';
+import type { ExperimentalMethod, Params, Response } from '../../params';
 
 export declare function nostrSignEvent(
     params: Params<NostrSignEvent & ExperimentalMethod>,
